requirements.txt
python-telegram-bot==21.3
anthropic==0.34.0
python-dotenv==1.0.1
